requirements.txt
feedparser==6.0.11
beautifulsoup4==4.12.3
Flask==3.1.0
gunicorn==23.0.0
psycopg2-binary==2.9.10
requests==2.32.3
